EG: avoided superfluous delete calls to the observers of Lucene; needed to reintroduce the return value of tracker.deleteDocId(). Also, refactored duplicate code around the call of lucene._deleteLucene into this method, removing its return value as that was no longer useful.
git-svn-id: file:///home/ozp/probeer-tj/meresco/meresco-components/workingsets/2.21.4-TUD/trunk@1782 83e7a582-4515-4195-8f39-222cbda6297e

+++ merescocomponents/facetindex/lucene.py
@@ -135,21 +135,18 @@ class LuceneIndex(Observable):
             if prevTxLuceneId != None:
                 docId = self._lucene2docId[prevTxLuceneId]
         if docId != None:
-            self._currentTracker.deleteDocId(docId)
-            self._commandQueue.append(FunctionCommand(self._delete, identifier=identifier))
-        return docId
+            deleted = self._currentTracker.deleteDocId(docId)
+            if deleted:
+                self._commandQueue.append(FunctionCommand(self._delete, identifier=identifier))
+                self.do.deleteDocument(docId=docId)
 
     def delete(self, identifier):
-        docId = self._luceneDelete(identifier)
-        if docId != None:
-            self.do.deleteDocument(docId=docId)
+        self._luceneDelete(identifier)
 
     def addDocument(self, luceneDocument=None):
         try:
             luceneDocument.validate()
-            oldDocId = self._luceneDelete(luceneDocument.identifier)
-            if oldDocId != None:
-                self.do.deleteDocument(docId=oldDocId)
+            self._luceneDelete(luceneDocument.identifier)
             docId = self._currentTracker.next()
             luceneDocument.docId = docId
             self._commandQueue.append(FunctionCommand(self._add, identifier=luceneDocument.identifier, document=luceneDocument))
